Update wgpu URL and bump Kotlin and wgpu versions to v24.0.3.1 (#51)
* Update wgpu URL and bump Kotlin and wgpu versions

Switched the wgpu.base.url to the gfx-rs source for consistency. Updated Kotlin to version 2.1.20 and wgpu to version v24.0.3.1 to include the latest improvements and fixes. These changes aim to enhance stability and maintain compatibility.

* Update AGP version to 8.7.3 in libs.versions.toml

Downgraded AGP from 8.9.0 to 8.7.3 for compatibility reasons. This ensures stability with current project dependencies and avoids potential build issues.

* Set WGPU library version in path resolution

Added a `version` constant to LibraryLoader and updated the library file path construction to include the version. This ensures the correct versioned library is resolved during loading.